GH#28684: Deduplicate GUI job refresh polling (#28686)
